Add Wikibase repoSiteName setting for client
Putting this outside of the normal Wikibase config,
so we can be absolutely sure it's available during
localisation cache rebuild.

The setting specifies a message key (set in Wikimedia Messages),
which is used by the WBREPONAME magic word (in various messages).

This setting does have a default that is set to 'Wikidata',
but we are setting it here to allow the name "Wikidata" to
be translatable (e.g. into Hebrew).

Change-Id: Iaa99d21627714cd3e9df8a842e9bfccad46c0763