Use the BetaFeatures whitelist for production to avoid accidental deploys
This uses the whitelist functionality in BetaFeatures (added in Ia19abd2403)
to actively manage the BetaFeature registrants live in production, avoiding
the accidental or poorly co-ordinated deploys that have hit us a few times
of late. It also uses the opportunity to actively manage the "6 month rule".

Beta Labs is left without a whitelist so that new BetaFeatures can be added
there for testing with the minimum of bureacracy and effort.

Change-Id: Ic04c7c8add3c047f8aaaa3f1255f3530edb0659f