Don't load as-yet-incompatible extensions when running under HHVM
This patch defines a global boolean $hhvm, set to true when running under HHVM
and false otherwise. Its value is used to prevent extensions that are
incompatible with HHVM from loading. At the moment, the extensions are
Scribunto (Tim is working on HHVM compatibility) and Wikibase (which should be
HHVM-compatible, but is throwing an error*).

HHVM is currently not enabled anywhere, but a roll-out to the beta cluster is
in the works.

* "Exception from line 49 of SettingsArray.php: Attempt to get non-existing
  setting 'defaultClientStore'."

Change-Id: I2d0e677b932914f025271c0295342c897bf08366