Update patch set 1
Patch Set 1:

The quickest way to persuade me to +2 this is to provide a clear explanation of why the need for better Hebrew support has to be met by enabling automatic font loading for a large number of scripts. I've asked that here, on-wiki, on Bugzilla, and via e-mail, and haven't gotten an answer yet.

Patch-set: 1