Modify wgAddGroups, wgRemoveGroups on brwikimedia
Allowing sysops to add users to, and remove them
from, the 'autopatrolled' and 'confirmed' user groups
in addition to 'translationadmin'.

Also allowing bureaucrats to remove users from the
'sysop' and 'bureaucrat' user groups.

See bug for URL with on-wiki consensus for this change
agreed upon by all current representatives of WMBR.

Bug: 63345
Change-Id: Ia65db6c432845cea715ef35bb8d1d41939ea2329