Add Musées de la Haute-Saône to wgCopyUploadsDomains
This patch changes in the whitelist $wgCopyUploadsDomains the domain
of the production web-server of the Musées de la Haute-Saône (France)
(so as to enable mass collection uploads on Wikimedia Commons
using the GlamWikiToolset).

It appeared whitelisting *.musees.cg70.fr was not good for
allowing uploads from musees.cg70.fr.

Follows-up on I0c60ef6d & I749f5fbc

Bug: 63449
Change-Id: Ifec1ae8d16482823bfc58f499669a52b5c48129c