Update patch set 1
Patch Set 1:

Where exactly do the "wmgMFRemovableClasses => extracts" get used currently? 
Is it safe/advisable for Hovercards to use this as a collection area, for "elements to remove"?
Could NavpopupsV2.0 use it for the same purpose in the future, too?
(Sorry, if I'm missing obvious documentation, somewhere. :)

Patch-set: 1
Label: Code-Review=0