Update patch set 2
Patch Set 2:

(2 comments)

Inline comments to Aaron.

To Faidon: Yeah, can totally use a different username/password/key, and that makes sense. As far as how much space: it's basically the size of the index with no replicas. Ballpark is a bit over 1TB to start (we're at about 4TB of disk utilization with all replicas in production at the moment).

Patch-set: 2