Update patch set 3
Patch Set 3:

We discussed it some more with Chad on IRC. Based on the size of the dataset (~1TB), let's wait until the new Swift servers to be fully weighted in the cluster (they are 66% as of today).

Additionally, because of considerations regarding ElasticSearch's transfer rate, we talked about saving one index at a time, starting with the smaller ones, and ramping up as we go.

Patch-set: 3
Reviewer: Faidon Liambotis <192@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0