Update patch set 2
Patch Set 2: Code-Review+1

The CPU hit we saw with the first patch I wrote to implement this on the MW side was due to parsing all ~180 addresses in the list as CIDR ranges.  This patch reduces that count to 22 CIDR ranges (12% of prior). I would not be surprised to see a small increase in CPU utilization as a result of this change but it shouldn't be anywhere near the spike that we saw previously.

Patch-set: 2
Reviewer: BryanDavis <1001@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1