Add SVG logos for nine Wikibooks wikis
This patch adds nine PNG thumbnails of SVG logos drawn
by Google Code-In students for the following Wikibooks
projects:

* azwikibooks (Azerbaijani);
* cywikibooks (Welsh — newly localised!);
* eowikibooks (Esperanto)
* etwikibooks (Estonian)
* euwikibooks (Basque — newly localised!);
* iewikibooks (Interlingua — newly localised!);
* mkwikibooks (Macedonian — newly localised!);
* ocwikibooks (Occitan);
* slwikibooks (Slovenian).

No community consensus was sought since the logos are
as close to 1:1 replicas as possible; see bug for
more information on that.

I decided /not to/ submit a new logo for the Ukrainian
Wikibooks as it significantly differs from the current
logo (by using a different, freely-licenced, font);
will do that in a patch of its own when/if local
community agrees for the change.

The logos are already protected on Commons against
re-uploads and name changes; I've done that myself.

Attribution notice:

All logos are released under the terms of the Creative
Commons Attribution-ShareAlike 3.0 licence (CC BY-SA).

* Azerbaijani Wikibooks logo by Natasha Murashkina;
* Welsh Wikibooks logo by Geoffrey Mon;
* Esperanto Wikibooks logo by Geoffrey Mon;
* Estonian Wikibooks logo by Natasha Murashkina;
* Basque Wikibooks logo by Geoffrey Mon;
* Interlingua Wikibooks logo by Yuan Siang;
* Macedonian Wikibooks logo by Geoffrey Mon;
* Occitan Wikibooks logo by Yuan Siang;
* Slovenian Wikibooks logo by Natasha Murashkina.

Bug: 52019
Change-Id: I8e602fe89de672f1922a2784e357845d62273b94