Configure git submodules to rebase on update
When performing an initial git clone of a new MediaWiki branch, add
a step to set the submodule.<name>.update key in .git/config to
'rebase'. "Defines what to do when the submodule is updated by the
superproject. [...] If rebase, the current branch of the submodule will
be rebased onto the commit specified in the superproject." [0]

It is hoped that this setting will make managing locally applied
security patches for extensions easier.

[0]: https://www.kernel.org/pub/software/scm/git/docs/gitmodules.html

Bug: 61562
Change-Id: I06a82003b37df831b6b146d8ea46980ee6520a04