Update patch set 5
Patch Set 5:

There doesn't appear to be any bug for the request, so 
    I'll just leave it as it is.

Can you explain, please? Does it mean you do not merge it? But why including this in the commit message?

BTW, overriding all the patches wasn't really intentional. I was downloading refs/changes/68/131968/2 and would have assumed gerrit to be smart enough to warn me if other users did some work in the meanwhile.

Patch-set: 5
Label: Verified=0