Update patch set 3
Patch Set 3: Code-Review+1

I'm inclined to think that Wikipedia:File_Upload_Wizard should be updated if not just replaced with [[mw:Extension:UploadWizard]] (which hopefully doesn't have the problems), rather than dumping people back on Special:Upload with the problems that Wikipedia:Upload was created to fix in the first place.

But let's not let the perfect be the enemy of the good here, and the on-wiki discussion seems to have a minimum of consensus. This can always be undone once the fancier things are fixed.

Patch-set: 3