Throttle GWToolset uploads
GWToolset can cause system instability by filling pages like
Special:NewFiles with large files which are expensive to thumbnail.
To avoid this, we are temporarily adding extra throttling.
The goal is to keep the ratio of GWToolset images below 10% on
NewFiles and co. That means:
* these pages default to the last 50 images, so GWT should not upload
  more than 5 images at the same time
* based on random checks of Special:Logs, there are about 240 uploads
  per hour in the slowest part of the day, so GWT's upload speed
  should not go over 10% of that.

We are overriding the default thread count, but not the max thread count
(which is 20), which means GWToolset users could override this limit
and upload up to 4x faster. There is no reason at the moment to prevent
that; it might be useful for uploading smaller files.

This uses a floating-point value for $wgJobBackoffThrottling, which
depends on I2e57d61fd67b97fbd593274b31e1bfada8f522f5

Change-Id: If6cc7e766f9b560db3b8fa385877c778d8436d15