Update patch set 1
Patch Set 1:

Do they necessarily know how expensive it will be to thumbnail the files, though? (File sizes are part of that, but not the only factor; file format, compression method used etc. can matter.) A large image collection can have a bunch of large files with similar names, amongst lots of smaller files; there is no reason the uploader would realize that.

At any rate, $wgJobBackoffThrottling seems superior to GWToolset's own throttling, which is per-user (that does not make sense from a performance standpoint). So even if we want to have a relatively permissive throttling, it would be better to use $wgJobBackoffThrottling for it.

If we want to have different limits for large and small files, defining multiple job types seems like a better approach to me.

Patch-set: 1