Update patch set 3
Patch Set 3:

Gergo, I don't think that applying a default to "large" wikis is a good idea. You keep looking at this issue as if a large wiki producing a lot of EL data is dangerous, when in fact the Analytics team has clearly stated that the EventLogging database could easily take that load and then some. This isn't a situation where we need to be paranoid about taking something down. Since Friday you've been looking at this as if it's a server-side emergency, but it's not. We've just been asked to dedicate less resources to Media Viewer logging, that's all.

Thus, we can focus on avoiding to lose valuable launch stats again. And that calls for an entirely different approach: let it all (or most of it) through and then adjust the resource consumption according to the amount we're really getting. The only thing that broke since last week is our stats, not Eventlogging, and we can't even know how the Commons launch went. Let's not do that again. And the existing wiki groups don't map well to the kind of per-wiki/metric tweaking we need to do here.

Patch-set: 3