Change banned Elasticsearch plugins
The hebrew plugin is a bit unstable - it'll crash some searches.
The unicode plugin should just be better.  Hopefully.  We'll deploy it in
phases by doing in place reindexes in this order:
group0
hebrew wikis

Then we'll wait a few days for soaking, then do group1, soak a few days,
then group2.

Change-Id: I19b5fe468abf4de56b5e446b4606d60b97afb729