Update patch set 1
Patch Set 1:

The only thing possibly worrying about this is that the ./php symlink will now be absolute rather than relative. I don't think this should be a problem however as I believe that the symlink is only used on the apaches and not needed on tin for any of the /a/common preparation scripts or scap scripts.

Patch-set: 1