Update patch set 3
Patch Set 3:

I think we'll have to do this in two stages: stage 1 would add the new variables but not remove the definition of wgMediaViewerSamplingFactor for 1.24wmf4 wikis, while stage 2 (to be merged after today's train deploy) would remove that obsolete variable once 1.24wmf4 is gone and I3966bf5d is backported.

I'll ping Reedy to double check that.

Patch-set: 3