Update patch set 1
Patch Set 1:

$wgJobBackoffThrottling['gwtoolsetUploadMetadataJob'] = 5 / 3600

will cause any successfully finishing gwtoolsetUploadMetadataJob to block the queue from picking another gwtoolsetUploadMetadataJob for 3600 / 5 seconds.

This patch is based on the assumption that the condition that brings down the scalers is N large TIFFs being present at the same time in Special:NewFiles. So it does not matter whether the uploads are evenly spaced out or clustered; what matters is there should not be more than N uploaded files in the time it normally takes for Special:NewFiles to rotate. (The value of N is a guess - I have no idea what the actual limits of the scalers are.)

So 5 file upload jobs are created every 12 minutes, which is roughly the time in which Special:NewFiles is replaced. I assumed that the file upload process itself is fast, so when uploads are started with big gaps between them, they will not finish at the same time.

$mediafile_job_throttle_max was left intact to allow some leeway for users uploading small files.

Patch-set: 1