Update patch set 1
Patch Set 1:

Ic62984e0f4a761642b2bdd1bfa362301ed94c284, if merged, will limit the number of parallel expensive thumbnailing attempts directly, so that should take care of small clusters in upload job end times. The goal of this change would be to avoid the forming of large clusters: the other patch uses PoolCounter, which means the number of running thumbnailing attempts is guaranteed to be less than the limit we set, but calls which are above that limit result in waiting HTTP requests which eventually causes the scalers to run into apache max connection limits.

So this patch should ensure files are not uploaded faster than they can be scaled; the limits used should probably be rethought.

Patch-set: 1