Update patch set 2
Patch Set 2:

Sorry for the confusion.  The new idea came up with at the new bug report yesterday.  'wotd' is not bad, but I believe 'featuredwords' is better for avoiding confusion.

Could we change it to featuredwords in the extension as well?  I wouldn't mind waiting for another week or so for deployment.

Patch-set: 2