Update patch set 2
Patch Set 2:

Ok, then I'd revert this to "wotd". I cannot ignore 50+ translations on TWN. "featuredwords" may be revisited as an addition when other types of feeds are wanted on Wiktionary.  dewiktionary has Word of the week, but they don't seem to mind using "wotd" at the bugzilla request.

Patch-set: 2