Set $wgIncludejQueryMigrate = true; for all wikis
MediaWiki 1.24 upgrades jQuery to 1.11. For one release there
is an opt-in variable to being back legacy APIs via the
jQuery Migrate plugin. WMF definitely needs this as various
extensions still use them.

Keeping it disabled (like MediaWiki does by default) on test2wiki
so that we have at least one environment in production that
represents the state we're moving towards. This should make it
harder to miss things in need of migration (since they'd be
working fine on other wikis thanks to $wgIncludejQueryMigrate).

Bug: 44740
Change-Id: I9dac0dc6a8015d0d883e8f01fb14f0d137d34010