Update patch set 1
Patch Set 1:

Yeah, I'm actually a bit surprised it was merged (I was about to abandon it, frankly, since pagemoves and editing cannot be logged), but I trust Legoktm reviewed it sufficiently. Perhaps it would be possible to only log the account name and not the IP. Another problem might be that when the log is viewed, there is no record (unlike CU log). Admins definitely should not be allowed to view it, as they can just add .* to the title blacklist and collect list of all IPs and usernames.

Patch-set: 1