Switch all wikis to the experimental highlghter
The highlighter has been soaking on commons and enwiki for weeks and weeks
with no complaints.  Its time to convert everyone.

DEPLOYMENT: This might marginally slow down highlighting until the index
is rebuilt.  Since we plan to rebuild indexes while rolling out 1.24wmf10
any way we should time this with the Tuesday roll out for wmf10 so it
gets in before our rebuild then.

Change-Id: I2e7d065fc757f2cd89616879c4e63fe5a044e060