Update patch set 12
Patch Set 12: Code-Review+1

Config-wise, looks good now. I'm not 100% convinced on the added complexity, but the DRY benefit probably outweighs it.

As for ruwiki, I suspect the inheritance is semi-intentional. These extra groups that get upload = true probably want the standard reupload permissions to go with it.

Note that reupload-own is redundant to reupload, the former is never checked if the latter is allowed.

Patch-set: 12
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1