Update patch set 3
Patch Set 3:

Just to confirm this in a place that isn't easily missed in future patchsets:

Bryan said "It sounds like we should fix parsoid in beta so that it works for your use case rather than trying to tie beta to production."

Yes. Very much yes. The Beta Cluster should be as self-contained as possible. If we're relying on (different version) code in production we're not really testing.

Patch-set: 3
Reviewer: Greg Grossmeier <680@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0