Update patch set 3
Patch Set 3:

> 1) We need to use cxserver-beta for the initial beta feature launch unless we suddenly get space from production.

That wasn't the plan as far as I remember. Getting it running on the Beta Cluster is a pre-step before getting it running on the prod hardware. Not a replacement for prod hardware. We *routinely* fall over if someone mistakenly starts requesting the deployment-prep vms to do anything meaningful in production.

> 2) Beta labs is not sufficient for us, since it does not contain non-English Wikipedias with content. Hence the cxserver backend service is already configured to fetch content from production parsoid to get real articles.

This is a mute point given above. BUT, we should extend the Beta Cluster as well with more content.

> 3) Gabriel told us that parsoid-beta is not well maintained currently. I do not know if they can fix it easily.

The parsoid/VE team should do a better job of maintaining it, if that's true.

Patch-set: 3