Juggle shard counts for Cirrus
Two things:
* Lower common's file index to 5 shards to reflect that it is more like
a content index then a 'general' index - it is searched frequently and
we want the searches to be lower overhead.
* Lower the maximum number of shards for a general index from 20 to 10.
This should lower the overhead of searching them at the cost of more overhead
replicating them.  We can live with slower replication.

DEPLOYMENT: This will require an in place reindex of the effected wikis
to take effect.  We'll only actually do commons and let the others come
in whenever we next reindex them which should be safe enough.

Change-Id: Ia72699ff45bc7262ff19f1e503ac102eb8692cd9