Update patch set 5
Patch Set 5: Code-Review+1

I was worried that using $IP/cache (instead of our wmf-config value of wgCacheDirectory, which is in /tmp), might be unexpected as it is unused and perhaps not set up properly. However we use it for l10n cache as well.

Patch-set: 5
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1