Update patch set 5
Patch Set 5:

Using $IP/cache was deliberate as this is a location that scap can sync across the cluster. The git json files have been syncing on each full scap for quite some time, but this config change and the mw-core patch it was dependent on are needed to make the feature work on wiki.

Patch-set: 5
Reviewer: BryanDavis <1001@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0