Enable job queue to process notification in test/test2 wiki
Job queue was turned off for Echo notification due to a db
lag between the main application and job runner, this was
causing issue in welcome notification, approximately 3 out
of 100 welcome notifications had corrupted data

This fix was merged in here: https://gerrit.wikimedia.org/r/#/c/81888/

We never had a chance to re-enable and test this, but we
should not delay this especailly for notification events that
would send tons of notifications, eg, the Flow topic subscription.
When Flow gets popular and widely deployed, job queue should
be used to handle the notification.

The goal is to first enable it on test/test2 wiki, if it looks
fine then enable it on all wikis

Change-Id: Ib0a7e80769cfb24a98bb7b82003a4a6383e3a04e