Update patch set 2
Patch Set 2:

> Wiki is locked, data's still in the database. I don't see what there is to fix.

I meant adding the star ratings and vote counts to the wikitext itself so they will still be visible. However, the page HTML is archived in the Wayback Machine, so I think we can go ahead with this. If anyone complains, a steward should be able to fix the pages without having DB access.

Patch-set: 2