Move RelatedSites config to wgExtraInterlanguageLinkPrefixes
This prevents the interwiki site setup from showing up in the list of
languages in user preferences.

I have removed those sites which aren't on the interwiki map, as that
config was just a no-op.

DO NOT MERGE until Icd91c80a2f3 and Id6b1aac8e82 are merged!!

Bug: 41209
Bug: 68359
Change-Id: I4f353cfb2cc96b0777d22a74d185c44b2886300e