Wikimedia wikis configuration for $wgExtraSignatureNamespaces
Newly introduced configuration variable (in I7ccf1093), defines the
namespaces that are expected to require the usage of signatures, in
addition to talk pages. In practice, it controls whether a button for
inserting signatures will be present in the toolbar.

MediaWiki default is array() (no namespaces). Most Wikimedia wikis use
NS_PROJECT as an additional discussion namespace (citation: just check
the interwikis of any Village pump subpage). I have been told that
Meta and Wikimania wikis also use the main namespace for this.

Change-Id: Ia4a950b96118d3deb7f76466f20d825cb2d8f807