Cirrus all field rollout stage two
Start using the all field in beta and get it building in production on
group0.

DEPLOYMENT: This will do nothing in production until the Cirrus search
index is rebuilt for the effected wikis.  Once the index is rebuilt we
can expect slightly higher load from indexes and we'll be able to test
searches of the all field with a special url parameter.

Change-Id: Iba5d01f220d94d05d5e4c9048fd4c3c28621fbc8