Update patch set 1
Abandoned

I am assuming this is a mistake since a revert got proposed as https://gerrit.wikimedia.org/r/#/c/153318/ already.

If on the other hands it is a revert war. Gerrit is not the appropriate place to handle it.  It is much better to raise the issue on wikitech-l.

Patch-set: 1
Status: abandoned