Update patch set 1
Patch Set 1:

Anomie and Reedy,  I remember category collation was something we had to be careful with.  There is a script we have to run and it doesn't work with large wikis.  shwiki looks to be small enough for it though.

Patch-set: 1
Reviewer: Manybubbles <873@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0