Remove bits.wikimedia.org/robots.txt
Also robots-private.txt, since this is apparently the only thing that
uses it, now that the robots.txt files for the private wikis have been
deleted. Requested by Google for much the same reason as IA -- they want
to be able to crawl CSS pages. There's no real reason to deny robots in
this domain -- it's not like there are expensive things here.

Change-Id: I85dcec2a168b9b2d42678d1d9a0e314793c99e21