Update patch set 1
Patch Set 1:

@Steinsplitter: Even if superprotection were going to be removed today (which seems unlikely despite your assertion), in the absence of bug 69640 the removal would be accomplished by removing 'superprotect' from the array rather than by reverting the entire change adding it. And similarly, here I am undoing the DRY cleanup that caused bug 69640 without affecting superprotection.


As for your rude question about my pay, the short answer is "No."

Long answer: I was a volunteer developer for long before I accepted employment with the Foundation, this issue came to my attention (via ping of my volunteer account on enwiki) outside of my normal working hours, the patch was written and submitted outside of my normal working hours, and it's exactly the sort of thing I would have done as a volunteer developer. On the other hand, it's also exactly the sort of patch expected of me in my particular position at the Foundation. On the balance, I'd lean towards "no, I'm not being paid for this patch" since this bug is non-urgent enough that the patch could wait until Monday were I not inclined to fix it in my volunteer role.

Patch-set: 1
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0