Update patch set 2
Patch Set 2: Verified+1 Code-Review+1

I have removed all unrelated -1 votes which are related to super protect. This should be talked about on wikitech-l or wikimedia-l or wherever you want as long as it is NOT in Gerrit.  Period.

----------

As for the change the comment explains the sort order issue.  We could come up one day with a syntax to optionally prepend but that is overkill compared to copy pasting :D

The rights are all good so it is good to go.  I am no going to deploy that on a saturday night at 10pm though.

Thanks Brad.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+1