Update patch set 1
Patch Set 1: Code-Review-1

Beta Labs is meant to be very similar to production (basically, let me run what's going to be in production in ~1 week).

I suggest putting this somewhere else (e.g. mwui.wmflabs.org, which already exists), or gating it with a URL parameter.

We used a similar parameter (in core) for a single special page (Special:UserLogin(/signup)) in the past.  In this case, I think the gate could just be in the URL.

Patch-set: 1
Reviewer: Mattflaschen <519@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1