Update patch set 8
Patch Set 8: Code-Review+1

This looks good to me. The string "labswiki" only appears as a substring in 5 lines of deleted.dblist now and the tests (such as they are) pass. As soon as this is merged the contents of the labswiki database will need to be copied to the new deploymentwiki database. I'm not completely clear on what services deployment.wikimedia.beta.wmflabs.org provides in beta that will need to be tested.

Patch-set: 8
Reviewer: BryanDavis <1001@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1