Update Parsoid extension require path
The Parsoid extension currently provides a symlink from /Parsoid.php to
/php/Parsoid.php. This patch changes the require to use the symlink instead of
the actual PHP file.

This is in preparation for https://gerrit.wikimedia.org/r/#/c/156169/ landing,
which will move the entire extension code to the root of the parsoid extension
directory, and replace the symlink with the actual file.

Change-Id: I6b66bd6dd7be6fe41119325a946f22fa069eae2e