Update patch set 1
Patch Set 1:

Are we going to announce and encourage the community to evaluate where these aren't needed? After all, if we keep this silent and just substitute it everywhere, the separation wasn't of much use. I think frwiktionary/botadmin is a case where editinterface was probably granted to edit settings pages that the bot reads, not js/css affecting the site as a whole.

Patch-set: 1