Turn on job throttling for Cirrus template jobs
Cirrus jobs that react to template changes are already lower priority then
those that react to direct edits but they seem to come in bursts and can
cause load spikes on the Elasticsearch servers.  We've lowered the maximum
number of concurrent in flight updates on the Elasticsearch side and raised
the maximum queue depth and timeout but that is much less elegant then
slowing down the jobs at the source with $wgJobBackoffThrottling.

Change-Id: I5aeda7cf610fb7e5cad082aec2c7fc682848bee3