Update patch set 1
Patch Set 1:

The actual throttle value that I picked it totally arbitrary.  I'm not sure what it'll do in terms of slow down or speed up.  If we don't like it we can go lower (.25, .1, whatever) or higher (5, 10, whatever).

The trick is that each job queue executor process applies the number independently so adding more nodes would increase the speed if the number isn't lowered.

Patch-set: 1
Label: Verified=0