Allow import from any WMF project to any other
It's been a long time coming, but here it is.

The old wgImportSources configuration has been completely removed, as there
is no need for it. In its place, we have $wmgCommonlyUsedImportSources,
which is used to place $lang and en at the top of the list for convenience,
and $wmgAdditionalImportSources, which is used to add chapter wikis and
Wikimanias as import sources for meta.

Also adding a missing ) to the import user agent.

DO NOT MERGE until I442a9afadc75f42f63d8c9c32ee2920374ceb26d is merged.
Otherwise, users will get annoyed.

Bug: 15583
Change-Id: Ib325df7c76837e1936b320d2a2441f10405090cd