Lower throttle for Cirrus template update jobs
The first throttle level wasn't enough to start building up a backlog.  Lets
half it and see if that gets us a backlog.  The goal here is to get a
constant backlog so we're seeing something close to even distribution of
the jobs across time.  It'd be a bonus if we can get job deduplication out
of it but with only a few minutes of backlog that is unlikely.

Change-Id: If678aa012c64867ccfcdb31e70830a87bafa423d