Update patch set 2
Patch Set 2:

Where is the user while the delay is happening? are they still on the login page, waiting on a page reload or some liminal state between clicking on Login and the page loading. 

4 seconds feels long

facebook is ~2 
google is <1 (feels instant)

I'd like to get Steven's perspective as well.

Patch-set: 2