Update patch set 2
Patch Set 2:

SUL does some redirects, but visually what you actually see now is the login page still in front of you after you click the green "log in" button. The only thing going on during this delay that you can see is a browser spinner or the like. After that you are taken to the page you started on before pressing login.

The ~2 delay visually adds on to the existing delay in the same manner that it already exists now (just more time).

Patch-set: 2