Update patch set 2
Patch Set 2:

We don't currently have a component which displays progress in a button, we will have this component as part of mediawiki.ui So while I would normally say  yes, to that Dzhan said, lets just use a progress indicator in the login button, we don't have something ready to use yet. 

I'm not 100% comfortable giving the go ahead on this, If we can wait till HHVM and the indeterminate loading control is in place I'd feel better about this change. 

I'll defer to Steven.

Patch-set: 2