Update patch set 1
Patch Set 1: Code-Review-1

We use tabs for all PHP code in the MediaWiki ecosystem, this repo shouldn't be different.

I personally think tabs are superior because the tab character was literally designed for this purpose. Using tabs allows people to configure the tab width in their code editor, whereas this isn't possible with spaces. 

If we want to have a discussion about tabs vs spaces in MW PHP code, we can have that discussion, but let's not change it in just one repo. All MW PHP code should follow the same style, and if we want to change that (which I don't, but who knows, maybe I'm in the minority) then we should change it everywhere at the same time.

Patch-set: 1
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1