Update patch set 1
Patch Set 1:

commented in the bug as well:

looking at the URL, http://www.soumaya.com.mx/, everything is “protected” via a flash application. i can”t tell if the images reside at this domain or elsewhere, so whitelisting this domain may not be helpful.

do they have an example of a few image URLs they wish to upload? that would help identify the domain that needs to be whitelisted.

Patch-set: 1