Lower throttle on two cirrus jobs
This lowers the throttle on the job reacting to template changes again
because the current value still wasn't enought to cause a small backlog
to appear.

This introduces a throttle for the job the counts links.  The throttle is
set the same as the template change jobs because we have fewer count links
jobs to deal with then the template change jobs so it should be enough.
The goal of introducing this throttle isn't to create a backlog for
deduplication its to make sure we don't overwhelm elasticsearch with too
many of these at once.

Change-Id: I9ddbdf559371bc882ca2fd851ed79f3f7d3cb817