Update patch set 2
Patch Set 2:

I'm not sure how Gerrit handles the rendering of these things, but someone with the ability to test the patch on a retina device as a favicon should do so. Smaller file size is good, but we don't want it to look shitty for these displays.

Patch-set: 2
Reviewer: Jaredzimmerman <1697@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0