Change how GettingStarted Redis server IP is determined
Previously it used the same Redis server as sessionRedis.  This works
fine in production.  However it fails for Labs since it has its own
session-labs.php that (unlike production's session.php) does not use
Redis.

This allows explicitly specifying the Redis host name.  However, it
keeps the current behavior (using sessionRedis) as a fallback, so
production is not affected.

Bug: 59838
Change-Id: I93c151a5f6e6cd010c91ffd0c116fdcc4826cd42