HHVM: capture & log traces for catchable fatals.
This patch adds HHVMRequestInit.php, meant to be used as the value of
`hhvm.server.request_init_document` setting. As its name suggests, it will be
run at the beginning of each request, before any other PHP code is invoked.

For the moment, the purpose of this file is to set up an error handler for
catchable PHP fatals that logs traces to syslog. PHP5 cannot call user error
handlers with fatals (this is why we have the wmerrors extension), but HHVM
can.

Change-Id: I5ff22ac23f11cb71c348cad8c33ac6e3f49e9d8f