Experimentally enable vips for larger (>50MP) tiff files
This also enables sharpening for tiff files (Its been asked
for for quite a while, and I think it'd be less disruptive to
change both the renderer and this at the same time, instead of
2 separate rendering program changes)

Currently we don't render such files at all, so in terms of an
experiment its a good thing to start with, since worst case
scenario, it doesn't work (or gives bad results), it can't be
any worse than the current strategy of not even trying.

Should this be succesful, and if the community is on board, I
would like to eventually extend this to all files.

Note: To be clear, the condition here of minArea = 5e7 applies to
the original file. Thus if the tiff file is resized in multiple
stages, vips will either be used in both stages, or not at all.

Bug: 52045
Change-Id: I1bd4a283281f2a18b9955c37fd8c34600f4015e2