Update patch set 1
Patch Set 1:

gwicke says: Extension:Parsoid is configured "basically everywhere either flow or VE is enabled; pretty much everywhere these days".  Is this true?

If not we'll probably need to do some additional configuration on wikis where Extension:Collection is enabled but not Extension:Parsoid.

And we need to ensure that the Parsoid service knows how to locate all these new wikis.  The Parsoid configuration is here:
https://git.wikimedia.org/blob/mediawiki%2Fservices%2Fparsoid%2Fdeploy/38d44ada74affeabb9f5c2bea798cea7af2d722f/conf%2Fwmf%2Flocalsettings.js
and:
https://git.wikimedia.org/blob/mediawiki%2Fservices%2Fparsoid/master/lib%2Fmediawiki.ParsoidConfig.js#L12

Patch-set: 1