Split cirrus's pool counters
Split Cirrus's pool counters in half - one half for enwiki the other half
for everything else.  The idea is that if something goes wrong with enwiki
it won't cause too much trouble for other wikis.  This does have the side
effect of doubling the number of actions that can be in flight at any one
time.

Also enables fast regex searching on enwiki because enwiki is ready for it
and we may as well turn it on there while we're here.

Change-Id: Id6c7a27315b1f32ea442dec25cd87a5d811b0fa6