Update patch set 1
Patch Set 1:

The change is technically fine, but I would like people to confirm wikiportret.nl is actually suitable for whitelisting.

I have relied on dan-nl +1 so far, but would be nice to scale it to some kind of groups that could reviews the request in bugzilla, approve it and THEN have the configuration change proposed =)

Waiting for dan-nl to vote on this one meanwhile.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0