Update patch set 2
Patch Set 2:

I really think we should revert this. This was not discussed and Special:Random (linked on the sidebar) also now shows pages in Draft namespace and if the page is in that namespace it's not intended for readers.

Patch-set: 2
Reviewer: Glaisher <892@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0