Revert "enwiki: Add Draft: namespace to wgContentNamespaces"
I understand the motivation as explained to me in person for
making this change: there were problems with analytics on
page creation statistics.

$wgContentNamespaces seems like a reasonable thing to
change at first to help here. Except many things rely on it
* Special:Random on its default query (ie: the one from the sidebar)
* Special:ShortPages (probably lots of other special pages too)
* Cirrus requires reindexing (not urgent, but was going to be
painful for the 2nd time in a week)
* Default search queries now include draft results (debatable whether that's desired behavior).

I'm thinking that editing metrics need to be arrived at another
way than changing this /very functional/ setting. Before 
attempting this again there should be a bug and/or wikitech-l
discussion about the impact.

This reverts commit 580fc43d3251aa90e5d9e288ca88ced5e306dc79.

Change-Id: I46b151ff40a193e6e4bacae41d77ac8df9bfdd8d